feat: add the ability to group instances between node definitions
This feature allows for grouping instances between single or multiple
node definitions. Assigning the group is optional and if set will modify
the logger created for the node to include the group assigned. The
responses have been updated to include the group or when retrieving the
hosts the groups that the host has associated with it.

The responses have been updated to remove the use of the API definitions
that are generated from the openapi in order for future versioned
RESTful API to be implemented. All tests have been updated to use
generic map[string]interface{} defined values where applicable for
complete end-to-end testing.

The database was also updated to add a GetHostsByGroup method with
corresponding index that will open up the API for the versioned API.