feat: monitoring and observability using OpenTelemetry
This feature allows the enablement of OpenTelemetry monitoring and
observability. Traces/Spans and metrics are available for most
functionality pertaining to admin API, WebSocket(s), and in-memory
database usage. This will allow proper setting of monitoring and alerts
while testing your control plane to ensure stability through the
lifecycle of your application gateway.